requirements.txt
smolagents == 1.17.0
python-dotenv == 1.1.0
smolagents[litellm]
pyserial
requests